requirements.txt
django==4.1.10
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
django-redis==5.3.0
django-cacheops==7.0.1
celery==5.3.4
django-celery-results==2.5.1
flower==2.0.1
freezegun==1.2.2
boto3==1.28.51
fabric2==3.2.2
requests==2.31.0
flake8==6.1.0
flake8-import-order==0.18.2
psycopg2==2.9.8
pydantic==2.4.2
pydantic_core==2.10.1
gunicorn==21.2.0
botocore==1.31.51
PyJWT==1.7.1
cryptography==41.0.4
newrelic==9.6.0
Jinja2==3.1.3
openai==1.35.10
pillow==10.4.0
sentry-sdk==2.11.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
django-constance[redis]==4.1.2